Software/sandbox/AprilTag_PoseDetector/requirements.txt
opencv-contrib-python
numpy
PyYAML
Pillow
reportlab

